Done Merr

+++ database/migrations/2024_06_23_071818_create_menus_table.php
@@ -16,10 +16,10 @@ class CreateMenusTable extends Migration
         Schema::create('menus', function (Blueprint $table) {
             $table->id();
             $table->string('nama', 50);
-            $table->integer('energi');
-            $table->integer('protein');
-            $table->integer('lemak');
-            $table->integer('karbo');
+            $table->float('energi');
+            $table->float('protein');
+            $table->float('lemak');
+            $table->float('karbo');
             $table->text('bahan');
             $table->text('cara');
             $table->timestamps();
